server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
google-generativeai>=0.3.0
pandas>=1.3.0
numpy>=1.20.0
tabulate>=0.8.10 